Use typedefs whenever possible
