+++ cvc4/include/cvc4_solver.h
@@ -48,7 +48,7 @@ class CVC4Solver : public AbsSmtSolver
                  const Sort & idxsort,
                  const Sort & elemsort) const override;
   Sort make_sort(SortKind sk,
-                 const std::vector<Sort> & sorts,
+                 const SortVec & sorts,
                  const Sort & sort) const override;
   Term make_value(bool b) const override;
   Term make_value(int64_t i, const Sort & sort) const override;
@@ -64,7 +64,7 @@ class CVC4Solver : public AbsSmtSolver
                  const Term & t0,
                  const Term & t1,
                  const Term & t2) const override;
-  Term make_term(Op op, const std::vector<Term> & terms) const override;
+  Term make_term(Op op, const TermVec & terms) const override;
   void reset() override;
   void reset_assertions() override;
   bool has_symbol(const std::string name) const override;
