+++ msat/include/msat_solver.h
@@ -64,7 +64,7 @@ class MsatSolver : public AbsSmtSolver
                  const Sort & idxsort,
                  const Sort & elemsort) const override;
   Sort make_sort(SortKind sk,
-                 const std::vector<Sort> & sorts,
+                 const SortVec & sorts,
                  const Sort & sort) const override;
   Term make_value(bool b) const override;
   Term make_value(int64_t i, const Sort & sort) const override;
@@ -80,7 +80,7 @@ class MsatSolver : public AbsSmtSolver
                  const Term & t0,
                  const Term & t1,
                  const Term & t2) const override;
-  Term make_term(Op op, const std::vector<Term> & terms) const override;
+  Term make_term(Op op, const TermVec & terms) const override;
   void reset() override;
   void reset_assertions() override;
   bool has_symbol(const std::string name) const override;
