+++ tests/msat/msat-tests.cpp
@@ -60,8 +60,8 @@ int main()
   Term x_lower = s->make_term(ext30, x);
   Term x_ext = s->make_term(Op(Zero_Extend, 4), x_lower);
 
-  Sort funsort = s->make_sort(
-      FUNCTION, std::vector<Sort>{ x_lower->get_sort() }, x->get_sort());
+  Sort funsort =
+      s->make_sort(FUNCTION, SortVec{ x_lower->get_sort() }, x->get_sort());
   Term uf = s->make_term("f", funsort);
   Term uf_app = s->make_term(Apply, uf, x_lower);
   assert(uf_app->get_op() == Apply);
